Move Replay modal to dedicated folder
