+++ ui/apps/dashboard/src/gql/gql.ts
@@ -47,8 +47,6 @@ const documents = {
     "\n  query GetFnCancellations($after: String, $envSlug: String!, $fnSlug: String!) {\n    env: envBySlug(slug: $envSlug) {\n      fn: workflowBySlug(slug: $fnSlug) {\n        cancellations(after: $after) {\n          edges {\n            cursor\n            node {\n              createdAt\n              envID: environmentID\n              id\n              name\n              queuedAtMax\n              queuedAtMin\n            }\n          }\n          pageInfo {\n            hasNextPage\n          }\n        }\n      }\n    }\n  }\n": types.GetFnCancellationsDocument,
     "\n  mutation InvokeFunction($envID: UUID!, $data: Map, $functionSlug: String!, $user: Map) {\n    invokeFunction(envID: $envID, data: $data, functionSlug: $functionSlug, user: $user)\n  }\n": types.InvokeFunctionDocument,
     "\n  query GetHistoryItemOutput($envID: ID!, $functionID: ID!, $historyItemID: ULID!, $runID: ULID!) {\n    environment: workspace(id: $envID) {\n      function: workflow(id: $functionID) {\n        run(id: $runID) {\n          historyItemOutput(id: $historyItemID)\n        }\n      }\n    }\n  }\n": types.GetHistoryItemOutputDocument,
-    "\n  query GetReplayRunCounts($environmentID: ID!, $functionSlug: String!, $from: Time!, $to: Time!) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        replayCounts: replayCounts(from: $from, to: $to) {\n          completedCount\n          failedCount\n          cancelledCount\n          skippedPausedCount\n        }\n      }\n    }\n  }\n": types.GetReplayRunCountsDocument,
-    "\n  mutation CreateFunctionReplay(\n    $environmentID: UUID!\n    $functionID: UUID!\n    $name: String!\n    $fromRange: ULID!\n    $toRange: ULID!\n    $statuses: [ReplayRunStatus!]\n  ) {\n    createFunctionReplay(\n      input: {\n        workspaceID: $environmentID\n        workflowID: $functionID\n        name: $name\n        fromRange: $fromRange\n        toRange: $toRange\n        statusesV2: $statuses\n      }\n    ) {\n      id\n    }\n  }\n": types.CreateFunctionReplayDocument,
     "\n  query GetReplays($environmentID: ID!, $functionSlug: String!) {\n    environment: workspace(id: $environmentID) {\n      id\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        replays {\n          id\n          name\n          createdAt\n          endedAt\n          functionRunsScheduledCount\n        }\n      }\n    }\n  }\n": types.GetReplaysDocument,
     "\n  query GetFunctionPauseState($environmentID: ID!, $functionSlug: String!) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        isPaused\n      }\n    }\n  }\n": types.GetFunctionPauseStateDocument,
     "\n  mutation NewIngestKey($input: NewIngestKey!) {\n    key: createIngestKey(input: $input) {\n      id\n    }\n  }\n": types.NewIngestKeyDocument,
@@ -111,6 +109,8 @@ const documents = {
     "\n  mutation testReplication($input: CDCConnectionInput!, $envID: UUID!) {\n    cdcTestLogicalReplication(input: $input, envID: $envID) {\n      steps\n      error\n    }\n  }\n": types.TestReplicationDocument,
     "\n  mutation testAutoSetup($input: CDCConnectionInput!, $envID: UUID!) {\n    cdcAutoSetup(input: $input, envID: $envID) {\n      steps\n      error\n    }\n  }\n": types.TestAutoSetupDocument,
     "\n  mutation cdcDelete($envID: UUID!, $id: UUID!) {\n    cdcDelete(envID: $envID, id: $id) {\n      ids\n    }\n  }\n": types.CdcDeleteDocument,
+    "\n  query GetReplayRunCounts($environmentID: ID!, $functionSlug: String!, $from: Time!, $to: Time!) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        replayCounts: replayCounts(from: $from, to: $to) {\n          completedCount\n          failedCount\n          cancelledCount\n          skippedPausedCount\n        }\n      }\n    }\n  }\n": types.GetReplayRunCountsDocument,
+    "\n  mutation CreateFunctionReplay(\n    $environmentID: UUID!\n    $functionID: UUID!\n    $name: String!\n    $fromRange: ULID!\n    $toRange: ULID!\n    $statuses: [ReplayRunStatus!]\n  ) {\n    createFunctionReplay(\n      input: {\n        workspaceID: $environmentID\n        workflowID: $functionID\n        name: $name\n        fromRange: $fromRange\n        toRange: $toRange\n        statusesV2: $statuses\n      }\n    ) {\n      id\n    }\n  }\n": types.CreateFunctionReplayDocument,
     "\n  fragment TraceDetails on RunTraceSpan {\n    name\n    status\n    attempts\n    queuedAt\n    startedAt\n    endedAt\n    isRoot\n    outputID\n    stepID\n    spanID\n    stepOp\n    stepInfo {\n      __typename\n      ... on InvokeStepInfo {\n        triggeringEventID\n        functionID\n        timeout\n        returnEventID\n        runID\n        timedOut\n      }\n      ... on SleepStepInfo {\n        sleepUntil\n      }\n      ... on WaitForEventStepInfo {\n        eventName\n        expression\n        timeout\n        foundEventID\n        timedOut\n      }\n    }\n  }\n": types.TraceDetailsFragmentDoc,
     "\n  query GetRunTrace($envID: ID!, $runID: String!) {\n    workspace(id: $envID) {\n      run(runID: $runID) {\n        function {\n          app {\n            name\n            externalID\n          }\n          id\n          name\n          slug\n        }\n        trace {\n          ...TraceDetails\n          childrenSpans {\n            ...TraceDetails\n            childrenSpans {\n              ...TraceDetails\n            }\n          }\n        }\n        hasAI\n      }\n    }\n  }\n": types.GetRunTraceDocument,
     "\n  query TraceResult($envID: ID!, $traceID: String!) {\n    workspace(id: $envID) {\n      runTraceSpanOutputByID(outputID: $traceID) {\n        data\n        input\n        error {\n          message\n          name\n          stack\n        }\n      }\n    }\n  }\n": types.TraceResultDocument,
@@ -287,14 +287,6 @@ export function graphql(source: "\n  mutation InvokeFunction($envID: UUID!, $dat
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n  query GetHistoryItemOutput($envID: ID!, $functionID: ID!, $historyItemID: ULID!, $runID: ULID!) {\n    environment: workspace(id: $envID) {\n      function: workflow(id: $functionID) {\n        run(id: $runID) {\n          historyItemOutput(id: $historyItemID)\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetHistoryItemOutput($envID: ID!, $functionID: ID!, $historyItemID: ULID!, $runID: ULID!) {\n    environment: workspace(id: $envID) {\n      function: workflow(id: $functionID) {\n        run(id: $runID) {\n          historyItemOutput(id: $historyItemID)\n        }\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query GetReplayRunCounts($environmentID: ID!, $functionSlug: String!, $from: Time!, $to: Time!) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        replayCounts: replayCounts(from: $from, to: $to) {\n          completedCount\n          failedCount\n          cancelledCount\n          skippedPausedCount\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetReplayRunCounts($environmentID: ID!, $functionSlug: String!, $from: Time!, $to: Time!) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        replayCounts: replayCounts(from: $from, to: $to) {\n          completedCount\n          failedCount\n          cancelledCount\n          skippedPausedCount\n        }\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation CreateFunctionReplay(\n    $environmentID: UUID!\n    $functionID: UUID!\n    $name: String!\n    $fromRange: ULID!\n    $toRange: ULID!\n    $statuses: [ReplayRunStatus!]\n  ) {\n    createFunctionReplay(\n      input: {\n        workspaceID: $environmentID\n        workflowID: $functionID\n        name: $name\n        fromRange: $fromRange\n        toRange: $toRange\n        statusesV2: $statuses\n      }\n    ) {\n      id\n    }\n  }\n"): (typeof documents)["\n  mutation CreateFunctionReplay(\n    $environmentID: UUID!\n    $functionID: UUID!\n    $name: String!\n    $fromRange: ULID!\n    $toRange: ULID!\n    $statuses: [ReplayRunStatus!]\n  ) {\n    createFunctionReplay(\n      input: {\n        workspaceID: $environmentID\n        workflowID: $functionID\n        name: $name\n        fromRange: $fromRange\n        toRange: $toRange\n        statusesV2: $statuses\n      }\n    ) {\n      id\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
@@ -543,6 +535,14 @@ export function graphql(source: "\n  mutation testAutoSetup($input: CDCConnectio
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n  mutation cdcDelete($envID: UUID!, $id: UUID!) {\n    cdcDelete(envID: $envID, id: $id) {\n      ids\n    }\n  }\n"): (typeof documents)["\n  mutation cdcDelete($envID: UUID!, $id: UUID!) {\n    cdcDelete(envID: $envID, id: $id) {\n      ids\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query GetReplayRunCounts($environmentID: ID!, $functionSlug: String!, $from: Time!, $to: Time!) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        replayCounts: replayCounts(from: $from, to: $to) {\n          completedCount\n          failedCount\n          cancelledCount\n          skippedPausedCount\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetReplayRunCounts($environmentID: ID!, $functionSlug: String!, $from: Time!, $to: Time!) {\n    environment: workspace(id: $environmentID) {\n      function: workflowBySlug(slug: $functionSlug) {\n        id\n        replayCounts: replayCounts(from: $from, to: $to) {\n          completedCount\n          failedCount\n          cancelledCount\n          skippedPausedCount\n        }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  mutation CreateFunctionReplay(\n    $environmentID: UUID!\n    $functionID: UUID!\n    $name: String!\n    $fromRange: ULID!\n    $toRange: ULID!\n    $statuses: [ReplayRunStatus!]\n  ) {\n    createFunctionReplay(\n      input: {\n        workspaceID: $environmentID\n        workflowID: $functionID\n        name: $name\n        fromRange: $fromRange\n        toRange: $toRange\n        statusesV2: $statuses\n      }\n    ) {\n      id\n    }\n  }\n"): (typeof documents)["\n  mutation CreateFunctionReplay(\n    $environmentID: UUID!\n    $functionID: UUID!\n    $name: String!\n    $fromRange: ULID!\n    $toRange: ULID!\n    $statuses: [ReplayRunStatus!]\n  ) {\n    createFunctionReplay(\n      input: {\n        workspaceID: $environmentID\n        workflowID: $functionID\n        name: $name\n        fromRange: $fromRange\n        toRange: $toRange\n        statusesV2: $statuses\n      }\n    ) {\n      id\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
